Change double value to BigDecimal in budgetchanges

+++ src/main/java/com/kacwol/manageYourBudget/budgetplan/model/request/BudgetPlanRequestElement.java
@@ -4,11 +4,13 @@ import com.kacwol.manageYourBudget.category.model.CategoryDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 public class BudgetPlanRequestElement {
 
     private CategoryDto category;
 
-    private double value;
+    private BigDecimal value;
 }
