+++ src/main/java/com/kacwol/manageYourBudget/budgetplan/model/response/BudgetPlanElementResponseDto.java
@@ -4,15 +4,17 @@ import com.kacwol.manageYourBudget.category.model.CategoryDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class BudgetPlanElementResponseDto {
 
     private CategoryDto category;
 
-    private double plannedValue;
+    private BigDecimal plannedValue;
 
-    private double actualValue;
+    private BigDecimal actualValue;
 
     private boolean wasMet;
